Carriage and snorkel update
Updated the carriage to position the linear bearings along the sides of
the carriage, added hole in center. need to figure out some way to
integrate the hole size with the bearing positioning.
Added piping to the snorkel, will need to add some way for it to have
internal supports for printing.